feat: Product 조회 구현 및 리팩토링

+++ tenshop-api/src/main/java/tenshop/api/product/dto/CategoryRegisterParam.java
@@ -2,6 +2,8 @@ package tenshop.api.product.dto;
 
 public record CategoryRegisterParam (
 	String name,
-	Integer parentId
+	Long parentId
 ) {
 }
+
+
